# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRISM AI Decision Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CustomPersonaType = typeof CustomPersonaType[keyof typeof CustomPersonaType];


export const CustomPersonaType = {
  system: 'system',
  famous: 'famous',
  custom: 'custom',
} as const;

export interface CustomPersona {
  id: string;
  name: string;
  type?: CustomPersonaType;
  rules: string;
  /** @nullable */
  riskLevel?: number | null;
  /** @nullable */
  creativityLevel?: number | null;
  /** @nullable */
  aggressiveness?: number | null;
}

export interface AnalysisInput {
  /** @minLength 5 */
  question: string;
  /** @nullable */
  customPersonas?: CustomPersona[] | null;
}

export interface PersonaPerspectives {
  contrarian: string;
  expansionist: string;
  executionist: string;
  analyst: string;
}

export interface Analysis {
  id: number;
  question: string;
  personas: PersonaPerspectives;
  verdict: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  createdAt: string;
}

export interface AnalysisSummary {
  id: number;
  question: string;
  verdict: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  createdAt: string;
}

export interface BattleInput {
  /** @minLength 5 */
  question: string;
  /**
     * @minItems 2
     * @maxItems 4
     */
  personas: CustomPersona[];
}

export interface PersonaArgument {
  persona: string;
  stance: string;
  strengths: string[];
  weaknesses: string[];
}

export interface BattleResult {
  arguments: PersonaArgument[];
  winner: string;
  verdict: string;
}

export type DebateSummaryWinner = typeof DebateSummaryWinner[keyof typeof DebateSummaryWinner];


export const DebateSummaryWinner = {
  sideA: 'sideA',
  sideB: 'sideB',
  tie: 'tie',
} as const;

export interface DebateSummary {
  id: number;
  sideA: string;
  sideB: string;
  topic?: string;
  winner: DebateSummaryWinner;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  createdAt: string;
}

export type DebateRecordWinner = typeof DebateRecordWinner[keyof typeof DebateRecordWinner];


export const DebateRecordWinner = {
  sideA: 'sideA',
  sideB: 'sideB',
  tie: 'tie',
} as const;

export interface DebateSide {
  label: string;
  argument: string;
  keyPoints: string[];
}

export interface DebateRecord {
  id: number;
  sideA: string;
  sideB: string;
  topic?: string;
  winner: DebateRecordWinner;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  reasoning: string;
  sideAData: DebateSide;
  sideBData: DebateSide;
  createdAt: string;
}

export interface DebateInput {
  /** @minLength 1 */
  sideA: string;
  /** @minLength 1 */
  sideB: string;
  /** Optional context or framing for the debate */
  topic?: string;
}

export type DebateResultWinner = typeof DebateResultWinner[keyof typeof DebateResultWinner];


export const DebateResultWinner = {
  sideA: 'sideA',
  sideB: 'sideB',
  tie: 'tie',
} as const;

export interface DebateResult {
  id?: number;
  sideA: DebateSide;
  sideB: DebateSide;
  winner: DebateResultWinner;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  reasoning: string;
}

export interface ErrorResponse {
  error: string;
}