# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PRISM AI Decision Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /debates/{id}:
    get:
      operationId: getDebate
      tags: [debate]
      summary: Get a specific debate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debate found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /debate:
    post:
      operationId: runDebate2
      tags: [debate]
      summary: Run a factual debate between two opposing positions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateInput"
      responses:
        "200":
          description: Debate complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    DebateRecord:
      type: object
      required: [id, sideA, sideB, winner, confidence, reasoning, sideAData, sideBData, createdAt]
      properties:
        id:
          type: integer
        sideA:
          type: string
        sideB:
          type: string
        topic:
          type: string
        winner:
          type: string
          enum: [sideA, sideB, tie]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        reasoning:
          type: string
        sideAData:
          $ref: "#/components/schemas/DebateSide"
        sideBData:
          $ref: "#/components/schemas/DebateSide"
        createdAt:
          type: string
          format: date-time
    DebateInput:
      type: object
      required: [sideA, sideB]
      properties:
        sideA:
          type: string
          minLength: 1
        sideB:
          type: string
          minLength: 1
        topic:
          type: string
          description: Optional context or framing for the debate
    DebateSide:
      type: object
      required: [label, argument, keyPoints]
      properties:
        label:
          type: string
        argument:
          type: string
        keyPoints:
          type: array
          items:
            type: string
    DebateResult:
      type: object
      required: [sideA, sideB, winner, confidence, reasoning]
      properties:
        id:
          type: integer
        sideA:
          $ref: "#/components/schemas/DebateSide"
        sideB:
          $ref: "#/components/schemas/DebateSide"
        winner:
          type: string
          enum: [sideA, sideB, tie]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        reasoning:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string