# lib/api-zod/src/generated/types/debateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRISM AI Decision Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface DebateInput {
  /** @minLength 1 */
  sideA: string;
  /** @minLength 1 */
  sideB: string;
  /** Optional context or framing for the debate */
  topic?: string;
}