# lib/api-zod/src/generated/types/debateRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRISM AI Decision Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { DebateRecordWinner } from './debateRecordWinner';
import type { DebateSide } from './debateSide';

export interface DebateRecord {
  id: number;
  sideA: string;
  sideB: string;
  topic?: string;
  winner: DebateRecordWinner;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  reasoning: string;
  sideAData: DebateSide;
  sideBData: DebateSide;
  createdAt: Date;
}