# lib/api-zod/src/generated/types/debateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRISM AI Decision Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { DebateResultWinner } from './debateResultWinner';
import type { DebateSide } from './debateSide';

export interface DebateResult {
  id?: number;
  sideA: DebateSide;
  sideB: DebateSide;
  winner: DebateResultWinner;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  reasoning: string;
}