# lib/api-zod/src/generated/types/debateSide.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRISM AI Decision Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface DebateSide {
  label: string;
  argument: string;
  keyPoints: string[];
}